Firebase 設定反映＆初期化処理
