+++ pubspec.yaml
@@ -33,6 +33,8 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_inappwebview: ^4.0.0
+  firebase_core: '0.5.0'
+  firebase_messaging: '^7.0.0'
   cloud_firestore: ^0.14.0
 
 # For information on the generic Dart part of this file, see the
